Use a single app for row and column layout examples
