--- examples/layout/row/lib/main.dart
@@ -1,33 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart' show debugPaintSizeEnabled;
-
-void main() {
-  debugPaintSizeEnabled = true; // Remove to suppress visual layout
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter layout demo',
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('Flutter layout demo'),
-        ),
-        // #docregion body
-        body: Center(
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: [
-              Image.asset('images/pic1.jpg'),
-              Image.asset('images/pic2.jpg'),
-              Image.asset('images/pic3.jpg'),
-            ],
-          ),
-        ),
-        // #enddocregion body
-      ),
-    );
-  }
-}
